Update index.html
Makes way more sense than before, I added the functionality of sliding images from the center position before, but because parent div was positioned absolute, it wouldn't work in the way we wanted it to. Now because parent div has a maximum width, text-align of the child paragraph works as the way we wanted it to be.